DS4Updater 1.1.23
Support to skip downloading a language pack

+++ Updater2/MainWindow.xaml.cs
@@ -38,7 +38,7 @@ namespace Updater2
         protected XmlDocument m_Xdoc = new XmlDocument();
         protected string m_Profile = Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + "\\DS4Tool\\Profiles.xml";
         private int round = 1;
-
+        public bool downloadLang = true;
         public bool AdminNeeded()
         {
             try
@@ -165,41 +165,44 @@ namespace Updater2
         {
             sw.Reset();
             string lang = CultureInfo.CurrentCulture.ToString();
-            if (round == 1)
+            if (downloadLang && !lang.StartsWith("en"))
             {
-                Uri i = new Uri("http://ds4windows.com/Files/" + lang + ".zip");
-                sw.Start();
-                wc.DownloadFileAsync(i, exepath + "\\" + lang + ".zip");
-                round = 2; 
-                return;
-            }
-            if (round == 2)
-            {                
-                if (new FileInfo(exepath + "\\" + lang + ".zip").Length > 0)
+                if (round == 1)
                 {
-                    try { Directory.Delete(exepath + "\\" + lang); }
-                    catch { }
-                    try { ZipFile.ExtractToDirectory(exepath + "\\" + lang + ".zip", exepath); }
-                    catch (IOException) { }
-                }
-                else
-                {
-                    File.Delete(exepath + "\\" + lang + ".zip");
-                    Uri i = new Uri("http://ds4windows.com/Files/" + lang.Split('-')[0] + ".zip");
+                    Uri i = new Uri("http://ds4windows.com/Files/" + lang + ".zip");
                     sw.Start();
                     wc.DownloadFileAsync(i, exepath + "\\" + lang + ".zip");
-                    round = 3;
+                    round = 2;
                     return;
                 }
-            }
-            if (round == 3)
-            {
-                if (new FileInfo(exepath + "\\" + lang + ".zip").Length > 0)
+                if (round == 2)
                 {
-                    try { Directory.Delete(exepath + "\\" + lang); }
-                    catch { }
-                    try { ZipFile.ExtractToDirectory(exepath + "\\" + lang + ".zip", exepath); }
-                    catch (IOException) { }
+                    if (new FileInfo(exepath + "\\" + lang + ".zip").Length > 0)
+                    {
+                        try { Directory.Delete(exepath + "\\" + lang); }
+                        catch { }
+                        try { ZipFile.ExtractToDirectory(exepath + "\\" + lang + ".zip", exepath); }
+                        catch (IOException) { }
+                    }
+                    else
+                    {
+                        File.Delete(exepath + "\\" + lang + ".zip");
+                        Uri i = new Uri("http://ds4windows.com/Files/" + lang.Split('-')[0] + ".zip");
+                        sw.Start();
+                        wc.DownloadFileAsync(i, exepath + "\\" + lang + ".zip");
+                        round = 3;
+                        return;
+                    }
+                }
+                if (round == 3)
+                {
+                    if (new FileInfo(exepath + "\\" + lang + ".zip").Length > 0)
+                    {
+                        try { Directory.Delete(exepath + "\\" + lang); }
+                        catch { }
+                        try { ZipFile.ExtractToDirectory(exepath + "\\" + lang + ".zip", exepath); }
+                        catch (IOException) { }
+                    }
                 }
             }
             if (new FileInfo(exepath + "\\Update.zip").Length > 0)
